seperated team object files
